+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - GitHub issue feedback link.
+- Subtle “Mistake?” feedback links in proposal tooltips.
 
 ### Fixed